+++ native/src/java/datalevin/ni/Env.java
@@ -26,7 +26,8 @@ public class Env {
     /**
      * Factory method to create an instance
      */
-    public static Env create(String dir, long size, int maxReaders, int maxDbs) {
+    public static Env create(String dir, long size, int maxReaders, int maxDbs,
+                             int flags) {
 
         WordPointer ptr = UnmanagedMemory.malloc(SizeOf.get(WordPointer.class));
         Lib.checkRc(Lib.mdb_env_create(ptr));
@@ -38,9 +39,7 @@ public class Env {
         Lib.checkRc(Lib.mdb_env_set_maxdbs(env, maxDbs));
         Lib.checkRc(Lib.mdb_env_open(env,
                                      CTypeConversion.toCString(dir).get(),
-                                     Mask.set(Lib.MDB_NORDAHEAD(),
-                                              Lib.MDB_MAPASYNC(),
-                                              Lib.MDB_WRITEMAP()),
+                                     flags,
                                      0664));
 
         return new Env(ptr);